Execute the checks against a dataset to return a nested list.

Execute the checks against a dataset to return a nested list.

execute_checks(ds, checks)

Arguments

ds

The data.frame to be checked. Required.

checks

The list describing the check. Is the output of load_checks(). Required.

Examples

# Replace the two paths for your specific project
path_data    <- system.file("datasets/pt-event-biochemical.rds", package = "trawler")
path_checks  <- system.file("checks/checks-biochemical.yml", package = "trawler")

ds_pt_event  <- readr::read_rds(path_data)
checks       <- load_checks(path_checks)

execute_checks(ds_pt_event, checks)
#> $record_id_name
#> [1] "record_id"
#> 
#> $baseline_date_name
#> [1] "date_enrolled"
#> 
#> $record_id_link
#> [1] "<a href=\"https://bbmc.ouhsc.edu/redcap/redcap_v%s/DataEntry/index.php?pid=%s&arm=%s&id=%s&page=%s\" target=\"_blank\">%s</a>"
#> 
#> $github_file_prefix
#> [1] "https://github.com/OuhscBbmc/validator-1/blob/master"
#> 
#> $redcap_codebook
#> [1] "https://bbmc.ouhsc.edu/redcap/redcap_v10.5.1/Design/data_dictionary_codebook.php?pid=1612"
#> 
#> $smells
#> # A tibble: 12 × 12
#>    check_name     pass  description       priority debug bound_lower bound_upper
#>    <chr>          <lgl> <chr>                <int> <lgl>       <dbl>       <dbl>
#>  1 proportion_fe… TRUE  Proportion femal…        2 FALSE        0.25        0.75
#>  2 proportion_ma… TRUE  Proportion male …        2 FALSE        0.25        0.75
#>  3 mean_age       TRUE  Mean age of part…        2 FALSE       20          80   
#>  4 mean_serum_pr… FALSE Mean serum pre-a…        2 FALSE       32          39   
#>  5 mean_serum_cr… FALSE Mean serum creat…        2 FALSE        3          15   
#>  6 average_bmi_a… FALSE Average BMI is b…        2 FALSE       18          24   
#>  7 mean_serum_ch… TRUE  Average Choleste…        1 FALSE      100         140   
#>  8 dialysis_adeq… TRUE  Normal range for…        1 FALSE        1.2         5   
#>  9 average_serum… TRUE  Mean serum ferri…        1 FALSE      501        1200   
#> 10 nutritional_c… TRUE  Most patients ag…        2 FALSE        0.85        0.99
#> 11 definitive_di… TRUE  All study partic…        1 FALSE        1           1   
#> 12 normalized_pr… FALSE Average Normaliz…        1 FALSE        0           0.12
#> # … with 5 more variables: bounds_template <chr>, value_template <chr>,
#> #   equation <chr>, boundaries <chr>, value <dbl>
#> 
#> $smell_status
#> [1] "12 smells have been sniffed.  4 violation(s) were found."
#> 
#> $smells_inactive
#> # A tibble: 0 × 9
#> # … with 9 variables: check_name <chr>, description <chr>, priority <int>,
#> #   debug <lgl>, bound_lower <dbl>, bound_upper <dbl>, bounds_template <chr>,
#> #   value_template <chr>, equation <chr>
#> 
#> $rules
#> # A tibble: 14 × 8
#>    check_name   violation_count error_message   priority debug redcap_instrument
#>    <chr>                  <int> <chr>              <int> <lgl> <chr>            
#>  1 baseline_pr…              10 Serum pre-albu…        1 FALSE baseline_data    
#>  2 missing_ser…               3 Relevant nutri…        1 FALSE baseline_data    
#>  3 serum_preal…              15 Baseline preal…        2 FALSE baseline_data, v…
#>  4 serum_preal…               0 Baseline preal…        2 FALSE baseline_data, v…
#>  5 serum_preal…               0 Baseline preal…        2 FALSE baseline_data, v…
#>  6 serum_preal…               0 serum prealbum…        3 FALSE baseline_data, v…
#>  7 baseline_fi…               0 Serum prealbum…        3 FALSE baseline_data, v…
#>  8 daily_first…               0 In-addition to…        3 FALSE baseline_data, p…
#>  9 daily_prote…               0 npcr levels in…        3 FALSE baseline_data, p…
#> 10 hospitaliza…               2 Patient was ho…        1 FALSE completion_proje…
#> 11 optimal_dai…               7 Daily protein …        2 FALSE completion_proje…
#> 12 recommended…              10 NPCR values ar…        2 FALSE completion_data  
#> 13 npcr                       1 NPCR at comple…        2 FALSE completion_data  
#> 14 npcr_compar…               2 NPCR at comple…        3 FALSE completion_data  
#> # … with 2 more variables: passing_test <chr>, results <list>
#> 
#> $rule_status
#> [1] "14 rules were examined. 8 rule(s) had at least 1 violation. 50 total violation(s) were found."
#> 
#> $rules_inactive
#> # A tibble: 1 × 6
#>   check_name         error_message  priority debug redcap_instrume… passing_test
#>   <chr>              <chr>             <int> <lgl> <chr>            <chr>       
#> 1 pre_albumin_levels pre_albumin l…        3 FALSE visit_lab_date   "function (…
#> 
#> attr(,"class")
#> [1] "trawler_checks"

# saveRDS(execute_checks(ds_pt_event, checks), "inst/derived/biochemical.rds")