Load and parse checks in a yaml file, and return a list of useful objects.

Load and parse checks in a yaml file, and return a list of useful objects.

load_checks(path_checks)

Arguments

path_checks

The file path to the yaml file specifying the smells and rules. Required.

Examples

# Replace the path for your specific project
path_checks  <- system.file("checks/checks-biochemical.yml", package = "trawler")
load_checks(path_checks)
#> $record_id_name
#> [1] "record_id"
#> 
#> $baseline_date_name
#> [1] "date_enrolled"
#> 
#> $record_id_link
#> [1] "<a href=\"https://bbmc.ouhsc.edu/redcap/redcap_v%s/DataEntry/index.php?pid=%s&arm=%s&id=%s&page=%s\" target=\"_blank\">%s</a>"
#> 
#> $github_file_prefix
#> [1] "https://github.com/OuhscBbmc/validator-1/blob/master"
#> 
#> $redcap_codebook
#> [1] "https://bbmc.ouhsc.edu/redcap/redcap_v10.5.1/Design/data_dictionary_codebook.php?pid=1612"
#> 
#> $smells
#> # A tibble: 12 × 10
#>    check_name description priority debug bound_lower bound_upper bounds_template
#>    <chr>      <chr>          <int> <lgl>       <dbl>       <dbl> <chr>          
#>  1 proportio… Proportion…        2 FALSE        0.25        0.75 [%.2f, %.2f]   
#>  2 proportio… Proportion…        2 FALSE        0.25        0.75 [%.2f, %.2f]   
#>  3 mean_age   Mean age o…        2 FALSE       20          80    [%.0f, %.0f]   
#>  4 mean_seru… Mean serum…        2 FALSE       32          39    [%.0f, %.0f]   
#>  5 mean_seru… Mean serum…        2 FALSE        3          15    [%.0f, %.0f]   
#>  6 average_b… Average BM…        2 FALSE       18          24    [%.0f, %.0f]   
#>  7 mean_seru… Average Ch…        1 FALSE      100         140    [%.0f, %.0f]   
#>  8 dialysis_… Normal ran…        1 FALSE        1.2         5    [%.1f, %.1f]   
#>  9 average_s… Mean serum…        1 FALSE      501        1200    [%.0f, %.0f]   
#> 10 nutrition… Most patie…        2 FALSE        0.85        0.99 [%.2f, %.2f]   
#> 11 definitiv… All study …        1 FALSE        1           1    [%.2f, %.2f]   
#> 12 normalize… Average No…        1 FALSE        0           0.12 [%.2f, %.2f]   
#> # … with 3 more variables: value_template <chr>, equation <chr>,
#> #   boundaries <chr>
#> 
#> $smells_inactive
#> # A tibble: 0 × 9
#> # … with 9 variables: check_name <chr>, description <chr>, priority <int>,
#> #   debug <lgl>, bound_lower <dbl>, bound_upper <dbl>, bounds_template <chr>,
#> #   value_template <chr>, equation <chr>
#> 
#> $smell_names_md
#> [1] "12 smells have been defined:\n\n1. proportion_female_participants;\n2. proportion_male_participants;\n3. mean_age;\n4. mean_serum_prealbumin_level_at_baseline;\n5. mean_serum_creatinine_level_at_baseline;\n6. average_bmi_at_baseline;\n7. mean_serum_cholesterol_levels_at_baseline;\n8. dialysis_adequacy;\n9. average_serum_ferritin_levels_at_baseline;\n10. nutritional_counseling;\n11. definitive_diagnosis;\n12. normalized_protein_catabolic_rate;\n"
#> 
#> $rules
#> # A tibble: 14 × 6
#>    check_name    error_message     priority debug instrument    passing_test    
#>    <chr>         <chr>                <int> <lgl> <chr>         <chr>           
#>  1 baseline_pre… Serum pre-albumi…        1 FALSE baseline_data "function (d) {…
#>  2 missing_seru… Relevant nutriti…        1 FALSE baseline_data "function (d) {…
#>  3 serum_prealb… Baseline prealbu…        2 FALSE baseline_dat… "function (d) {…
#>  4 serum_prealb… Baseline prealbu…        2 FALSE baseline_dat… "function (d) {…
#>  5 serum_prealb… Baseline prealbu…        2 FALSE baseline_dat… "function (d) {…
#>  6 serum_prealb… serum prealbumin…        3 FALSE baseline_dat… "function (d) {…
#>  7 baseline_fir… Serum prealbumin…        3 FALSE baseline_dat… "function (d) {…
#>  8 daily_first_… In-addition to b…        3 FALSE baseline_dat… "function (d) {…
#>  9 daily_protei… npcr levels in s…        3 FALSE baseline_dat… "function (d) {…
#> 10 hospitalizat… Patient was hosp…        1 FALSE completion_p… "function (d) {…
#> 11 optimal_dail… Daily protein in…        2 FALSE completion_p… "function (d) {…
#> 12 recommended_… NPCR values are …        2 FALSE completion_d… "function (d) {…
#> 13 npcr          NPCR at completi…        2 FALSE completion_d… "function (d) {…
#> 14 npcr_compari… NPCR at completi…        3 FALSE completion_d… "function (d) {…
#> 
#> $rules_inactive
#> # A tibble: 1 × 6
#>   check_name         error_message   priority debug instrument  passing_test    
#>   <chr>              <chr>              <int> <lgl> <chr>       <chr>           
#> 1 pre_albumin_levels pre_albumin le…        3 FALSE visit_lab_… "function (d) {…
#> 
#> attr(,"class")
#> [1] "trawler_checks_definition"