Render (executed) checks to an html document with R Markdown

Render (executed) checks to an html document with R Markdown

render_rmarkdown(
  path_checks,
  directory_output,
  path_template = system.file("report-templates/rmarkdown-1/report-1.Rmd", package =
    "trawler")
)

Arguments

path_checks

The file path of the output of execute_checks(). Required.

directory_output

A writable directory to save the report outputs. The html file is the primary output. The markdown file is sometimes useful too. Required.

path_template

The rmd file, which serves as the report template. Defaults to the standard template built into the trawler package.

Examples

# Step 0: define paths.
#   So this package example executes on every machine, temp files are used.

# Replace the two paths for your specific project
path_data    <- system.file("datasets/pt-event-biochemical.rds", package = "trawler")
path_defs    <- system.file("checks/checks-biochemical.yml", package = "trawler")

# In your real code, this will probably in a static directory.
# If PHI or sensitive info is contained, use a secure location.
directory_output <- file.path(tempdir(check = TRUE), "trawler")
if (!dir.exists(directory_output)) dir.create(directory_output)
path_checks <- tempfile("trawler-checks-", fileext = ".rds", tmpdir = directory_output)

# Step 1: load the check definitions and the dataset to test
ds_pt_event  <- readr::read_rds(path_data)
checks       <- load_checks(path_defs)
#> Error in checkmate::assert_character(origin, any.missing = FALSE, len = 1,     pattern = "^(?:csv|REDCap)$"): argument "origin" is missing, with no default

# Step 2: execute the checks and save to an rds file
ds_pt_event |>
  execute_checks(checks) |>
  readr::write_rds(path_checks)
#> Error in eval(expr, envir, enclos): object 'checks' not found

# Step 3: render checks as an html report with R Markdown
render_rmarkdown(path_checks, directory_output)
#> Error in render_rmarkdown(path_checks, directory_output): Assertion on 'path_checks' failed: File does not exist: '/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpQ73IsT/trawler/trawler-checks-66121fe90d3.rds'.

# For the sake of this example, clean up temp files.
#    You probably do NOT want this line in your real code.
unlink(directory_output, recursive = TRUE)